module7-task1

+++ project/apps/blog/src/app/comment/comment.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import {BadRequestException, Injectable, NotFoundException} from '@nestjs/common';
 import {CommentRepository} from './comment.repository';
 import {PostService} from '../post/post.service';
 import {CommentEntity} from './comment.entity';
 import {CreateCommentDto} from './dto/create-comment.dto';
 import {CommentQuery} from './query/comment.query';
+import {COMMENT_NOT_FOUND, NOT_AUTHOR} from './comment.constant';
 
 @Injectable()
 export class CommentService {
@@ -18,8 +19,14 @@ export class CommentService {
     return this.commentRepository.create(commentEntity);
   }
 
-  public async delete(commentId: number) {
+  public async delete(commentId: number, userId: string) {
     const comment = await this.commentRepository.findById(commentId);
+    if (!comment) {
+      throw new NotFoundException(COMMENT_NOT_FOUND)
+    }
+    if (comment.userId !== userId) {
+      throw new BadRequestException(NOT_AUTHOR);
+    }
     await this.postService.decrementCommentCount(comment.postId);
     return this.commentRepository.destroy(commentId);
   }
